refactor: move chat files out of features
