+++ src/components/chat/ChatThread.tsx
@@ -1,4 +1,4 @@
-import type { ChatMessage } from '../types';
+import type { ChatMessage } from '../../types/chat';
 import ChatMessageBubble from './ChatMessageBubble';
 
 /**
