--- src/features/chat/index.ts
@@ -1,4 +1,0 @@
-// features/chat/index.ts
-// ✅ 원리: 외부에서 chat 기능을 가져갈 때 import 경로를 깔끔하게 만들기 위한 "출구 파일"
-
-export { default as ChatPageView } from './components/ChatPageView';
