+++ src/main.tsx
@@ -2,7 +2,7 @@ import { StrictMode } from 'react';
 import { createRoot } from 'react-dom/client';
 import App from './App';
 import './index.css';
-import { initTalkSocket } from './features/chat/socket/talkSocket';
+import { initTalkSocket } from './socket/talkSocket';
 
 const rootElement = document.getElementById('root')!;
