+++ src/pages/Chatpage.tsx
@@ -1,5 +1,5 @@
 // src/pages/ChatPage.tsx
-import { ChatPageView } from '../features/chat';
+import ChatPageView from '../components/chat/ChatPageView';
 
 export default function ChatPage() {
   return <ChatPageView />;
